wip: stash something unseen
