+++ frappe/__init__.py
@@ -407,127 +407,7 @@ def connect_replica() -> bool:
 	return True
 
 
-def get_site_config(sites_path: str | None = None, site_path: str | None = None) -> dict[str, Any]:
-	"""Return `site_config.json` combined with `sites/common_site_config.json`.
-	`site_config` is a set of site wide settings like database name, password, email etc."""
-	config = _dict()
-
-	sites_path = sites_path or getattr(local, "sites_path", None)
-	site_path = site_path or getattr(local, "site_path", None)
-
-	common_config = get_common_site_config(sites_path)
-
-	if sites_path:
-		config.update(common_config)
-
-	if site_path:
-		site_config = os.path.join(site_path, "site_config.json")
-		if os.path.exists(site_config):
-			try:
-				config.update(get_file_json(site_config))
-			except Exception as error:
-				click.secho(f"{local.site}/site_config.json is invalid", fg="red")
-				print(error)
-		elif local.site and not local.flags.new_site:
-			error_msg = f"{local.site} does not exist."
-			if common_config.developer_mode:
-				from frappe.utils import get_sites
-
-				all_sites = get_sites()
-				error_msg += "\n\nSites on this bench:\n"
-				error_msg += "\n".join(f"* {site}" for site in all_sites)
-
-			raise IncorrectSitePath(error_msg)
-
-	# Generalized env variable overrides and defaults
-	def db_default_ports(db_type):
-		if db_type == "mariadb":
-			from frappe.database.mariadb.database import MariaDBDatabase
-
-			return MariaDBDatabase.default_port
-		elif db_type == "postgres":
-			from frappe.database.postgres.database import PostgresDatabase
-
-			return PostgresDatabase.default_port
-
-		raise ValueError(f"Unsupported db_type={db_type}")
-
-	config["redis_queue"] = (
-		os.environ.get("FRAPPE_REDIS_QUEUE") or config.get("redis_queue") or "redis://127.0.0.1:11311"
-	)
-	config["redis_cache"] = (
-		os.environ.get("FRAPPE_REDIS_CACHE") or config.get("redis_cache") or "redis://127.0.0.1:13311"
-	)
-	config["db_type"] = os.environ.get("FRAPPE_DB_TYPE") or config.get("db_type") or "mariadb"
-	config["db_socket"] = os.environ.get("FRAPPE_DB_SOCKET") or config.get("db_socket")
-	config["db_host"] = os.environ.get("FRAPPE_DB_HOST") or config.get("db_host") or "127.0.0.1"
-	config["db_port"] = int(
-		os.environ.get("FRAPPE_DB_PORT") or config.get("db_port") or db_default_ports(config["db_type"])
-	)
-
-	# Set the user as database name if not set in config
-	config["db_user"] = os.environ.get("FRAPPE_DB_USER") or config.get("db_user") or config.get("db_name")
-
-	# vice versa for dbname if not defined
-	config["db_name"] = os.environ.get("FRAPPE_DB_NAME") or config.get("db_name") or config["db_user"]
-
-	# read password
-	config["db_password"] = os.environ.get("FRAPPE_DB_PASSWORD") or config.get("db_password")
-
-	# Allow externally extending the config with hooks
-	if extra_config := config.get("extra_config"):
-		if isinstance(extra_config, str):
-			extra_config = [extra_config]
-		for hook in extra_config:
-			try:
-				module, method = hook.rsplit(".", 1)
-				config |= getattr(importlib.import_module(module), method)()
-			except Exception:
-				print(f"Config hook {hook} failed")
-				traceback.print_exc()
-
-	return config
-
-
-def get_common_site_config(sites_path: str | None = None) -> dict[str, Any]:
-	"""Return common site config as dictionary.
-
-	This is useful for:
-	- checking configuration which should only be allowed in common site config
-	- When no site context is present and fallback is required.
-	"""
-	sites_path = sites_path or getattr(local, "sites_path", None)
-
-	common_site_config = os.path.join(sites_path, "common_site_config.json")
-	if os.path.exists(common_site_config):
-		try:
-			return _dict(get_file_json(common_site_config))
-		except Exception as error:
-			click.secho("common_site_config.json is invalid", fg="red")
-			print(error)
-	return _dict()
-
-
-def get_conf(site: str | None = None) -> dict[str, Any]:
-	if hasattr(local, "conf"):
-		return local.conf
-
-	# if no site, get from common_site_config.json
-	with init_site(site):
-		return local.conf
-
-
-class init_site:
-	def __init__(self, site=None):
-		"""If site is None, initialize it for empty site ('') to load common_site_config.json"""
-		self.site = site
-
-	def __enter__(self):
-		init(self.site)
-		return local
-
-	def __exit__(self, type, value, traceback):
-		destroy()
+from frappe.deprecation_dumpster import get_common_site_config, get_conf, get_site_config, init_site
 
 
 def destroy():
@@ -1568,40 +1448,14 @@ def unscrub(txt: str) -> str:
 	return txt.replace("_", " ").replace("-", " ").title()
 
 
-def get_module_path(module, *joins):
-	"""Get the path of the given module name.
-
-	:param module: Module name.
-	:param *joins: Join additional path elements using `os.path.join`."""
-	from frappe.modules.utils import get_module_app
-
-	app = get_module_app(module)
-	return get_pymodule_path(app + "." + scrub(module), *joins)
-
-
-def get_app_path(app_name, *joins):
-	"""Return path of given app.
-
-	:param app: App name.
-	:param *joins: Join additional path elements using `os.path.join`."""
-	return get_pymodule_path(app_name, *joins)
-
-
-def get_app_source_path(app_name, *joins):
-	"""Return source path of given app.
-
-	:param app: App name.
-	:param *joins: Join additional path elements using `os.path.join`."""
-	return get_app_path(app_name, "..", *joins)
-
-
-def get_site_path(*joins):
-	"""Return path of current site.
-
-	:param *joins: Join additional path elements using `os.path.join`."""
-	from os.path import join
-
-	return join(local.site_path, *joins)
+from frappe.deprecation_dumpster import (
+	get_all_apps,
+	get_app_path,
+	get_app_source_path,
+	get_module_list,
+	get_module_path,
+	get_site_path,
+)
 
 
 def get_pymodule_path(modulename, *joins):
@@ -1615,30 +1469,6 @@ def get_pymodule_path(modulename, *joins):
 		joins = [scrub(part) for part in joins]
 
 	return abspath(join(dirname(get_module(scrub(modulename)).__file__ or ""), *joins))
-
-
-def get_module_list(app_name):
-	"""Get list of modules for given all via `app/modules.txt`."""
-	return get_file_items(get_app_path(app_name, "modules.txt"))
-
-
-def get_all_apps(with_internal_apps=True, sites_path=None):
-	"""Get list of all apps via `sites/apps.txt`."""
-	if not sites_path:
-		sites_path = local.sites_path
-
-	apps = get_file_items(os.path.join(sites_path, "apps.txt"), raise_not_found=True)
-
-	if with_internal_apps:
-		for app in get_file_items(os.path.join(local.site_path, "apps.txt")):
-			if app not in apps:
-				apps.append(app)
-
-	if "frappe" in apps:
-		apps.remove("frappe")
-	apps.insert(0, "frappe")
-
-	return apps
 
 
 @request_cache
